Add resource dependency helpers with tests
- Add resource module with helper to determine resources required by enabled metrics
- Add helper to identify ready-to-use resources and those requiring preparation
- Add resource_test module with tests for both helpers
- Delete resource dependency related code from config and config test modules
- Move qc_config_factory fixture to new conftest.py in quality_control test package
